# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "GuessWat",
        "version": "1.0"
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/word/translate":{
            "post":{
                "operationId": "translate",
                "summary": "Api for translating the word from source language to target language",
                "description": "Api for translating the word from source language to target language.\n",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {   "description": "token",
                    "in": "header",
                    "name": "token",
                    "required": true},
                    {   "description": "The word which is to be translated",
                        "in": "body",
                        "name": "word",
                        "required": true},
                    {   "description": "Type of the word to be translated",
                    "in": "body",
                    "name": "type",
                    "required": true},
                    {   "description": "Source language code of the word",
                        "in": "body",
                        "name": "origin",
                        "required": true},
                    {   "description": "Target language code where the word will be translated",
                        "in": "body",
                        "name": "target",
                        "required": true}
                    ],
                "responses": {
                    "200": {
                        "description": "200 response",
                        "examples": {
                            "application/json": "The json for the user"
                        }
                    }
                }
            }
        },
...
    }
}